Changed early handling of member accesses so that to create a parsed statement like "x.y = z" one simply has to write:
  new CallExpr("=", new CallExpr(".", x, y), z)

where x and z are symbols and y is a string literal.

The primitives PRIMITVE_GET_MEMBER and PRIMITIVE_SET_MEMBER are only
introduced in setters, getters, and constructors, and specificially
the compiler generated codes.  User-defined constructors are specially
treated in a wrapped function, i.e., the user part is put in a
function of its own and called.  User-defined getters and setters
could eventually be handled the same way should we decide to support
them.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5086 3a8e244f-b0f2-452b-bcba-4c88e055c3ca