Implemented folding of parameters as a part of normalization.  Some binary operators are now folded and some constants are now propagated. A constant value is only propagated if it is a primitive literal and it is stored in a variable that the user called a param or a const.
Added a Vec<SymExpr*>* to the Symbol class which is filled in by the
utility routine compute_sym_uses(BaseAST* base) which determines all
of the uses (including defs) of a Symbol.  This is used to propagate
literals and to test constness.

This constant folding uses the folding designed for the interpreter.

Turned off folding of string concatentation of string literals which
is not yet handled.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5112 3a8e244f-b0f2-452b-bcba-4c88e055c3ca