Made normalize work when it is called more than once.  As of this checkin, it is now called twice (like cleanup, for testing purposes). Instantiated functions are now normalized.
Added a check to make sure the same function or the same type is not
instantiated more than instantiation-limit number of times.  This
limit can be set at the command line.  The default is currently 12.

Added a test of recursive instantiation on a function and a test of
infinite instantiation (that gives an error).

Added USR_STOP macro which stops the compiler if any USR_FATAL_CONT
errors have been encountered.  Otherwise, the compiler stops at the
end of the current pass.

Added USR_PRINT macro which is used to just print out a message to the
user.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5130 3a8e244f-b0f2-452b-bcba-4c88e055c3ca