Added support for unknown number of variable arguments as in
  function foo(i : integer ...?num) {

where num is a param set to the number of i's that are passed in to
this routine.  It can be named anything.

Moved creation of tuple types after normalization to
buildDefaultFunctions so that we can create all that we need at that
point.  We have a little difficulty creating new tuples because we put
them in the _chpl_htuple module and then need to renormalize that
module after building new ones.

Used variable argument functions to implement domains in a completely
rank-independent way.  Domains of arbitrary rank are now supported.
Removed future for 5D domains.

Our handling of tuples of size 1 as in 1*integer is a bit weird in
that we transform this into an integer.  I think we should leave it as
a 1-tuple and force it to be indexed, making the code more consistent
in the case where rank is 1, and the user wrote rank*integer.  In the
case of variable arguments, this would allow the user to always index
into the variable argument.

Cleaned up domain code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5369 3a8e244f-b0f2-452b-bcba-4c88e055c3ca