Cleanup of inlining so that it inlines the statements in the function, not the BlockStmt body, resulting in cleaner code.  Also removed an extra temporary for the return value now that the function is normalized.
Added Expr* FnSymbol::retExpr to store the return expression of a
function.  It is removed during normalization.  It takes the place of
using DefExpr::exprType for FnSymbol.  This helped fix a bug in which
the return expression was not in the proper scope.

Simplified iterator reconstruction.  An iterator now requires a return
type expression.  This is temporary until we have a normalization of
iterators.

Fixed construction of CondStmt to remove an extra BlockStmt.

Removed "typeof" function.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5385 3a8e244f-b0f2-452b-bcba-4c88e055c3ca