Improved implementation of _ddata class, moving it out of the prelude and putting it in _chpl_base.  Codegen is now making use of the same primitives that analysis and the interpreter are using.  I had difficulty with initializing _ddata of records because we can't just make each vector element be the representative element.  The interpreter must call the default constructor for records with no arguments?  I added code to reassign a new record to each element in the _chpl_base module.
Fixed a bug in which a field with a complicated type resulted in an
odd normalized form.  Added a test case to capture this.

Improved the arithmetic array implementation greatly, adding support
for arrays of arbitrary rank and not limiting their size to 128
elements.

We need to consider how arrays should be written and how they should
be stored.  We currently store and traverse them column major, but we
print them row major.  In MatLab, a 1D array is printed as a column,
and in that case, column major makes more sense to me.  I have a small
preference to print a 1D array as a column.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5496 3a8e244f-b0f2-452b-bcba-4c88e055c3ca