Modified expand such that if a single argument is given it is expanded this amount in all dimensions.
Added _aseq methods for interior,
exterior, and expand to simplify the
corresponding domain functions, with
this code as well as the _aseq translate
method placed in the _chpl_adomain.chpl
file.

Added error checking for translate,
interior, exterior, and expand to halt
if an incorrect number of arguments are
given.

Added more tests and updated some test
output.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5513 3a8e244f-b0f2-452b-bcba-4c88e055c3ca