Added basic support for indefinite domains and arrays in a new module _chpl_indefinite.  The indefinite domains support an add method to add new indexes.  A remove method is not yet supported.  This initial implementation uses quadratic probing and roughly doubles the prime size of the hash table when it becomes half full.  Iteration over indefinite domains is not yet supported.  Limited support includes the add method and declaring indefinite arrays over indefinite domains. Valid index types currently only include integer and string.  The current hash function is pathetic and requires some sort of primitive to access the bytes that make up primitive types.
Indefinite arrays are currently limited to 128 elements.  Limited
support includes indexing with the index type of the indefinite
domain.

Lost support for domains as fields in classes.  Domains and arrays are
now handled the same way and require something extra because the exact
type is now determined based on function resolution.

Moved functions common to arithmetic and indefinite arrays into
_chpl_base module.

Determined that 'or' and 'and' do not short-circuit.  This needs to be
handled by normalization.

Updated regressions file to show current regressions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5558 3a8e244f-b0f2-452b-bcba-4c88e055c3ca