Added bool FnSymbol::hasVarArgs which is true if any of the formal arguments in the function are unexpanded variable arguments.
Fixed the interpreter so that it may overinstantiate a generic
function that has variable arguments, thus fixing

  test/functions/deitz/varargs/test_varargs7.chpl

in the interpreter.

Added ability to index into an arithmetic domain with a dimension
number to determine the extent of that dimension.  For example, to
iterate over the first dimension of domain D, write

  forall i in D(1) ...


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5589 3a8e244f-b0f2-452b-bcba-4c88e055c3ca