Added support for instantiating formal argument types during normalization to allow for proper dispatch.  Removed future for test/types/seq/deitz/test_seq_arg.chpl.  This change also gets 1D arrays working through the interpreter.
Moved construction of tuple types to the cleanup pass and made it so
that we now create lots of tuple types right up front.  This slows us
down, though I am optimizing some parts of the slowdown.  We should
discuss the tuple story at our next meeting.

Futurized one test that used a field in a class that is an
instantiated type to deal with this later.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5635 3a8e244f-b0f2-452b-bcba-4c88e055c3ca