Added support for
  FOR PARAM identifier IN expr..expr stmts

which creates a loop that is unrolled.  It is currently implemented
via recursive instantiation, inlining, and nested functions.

At some point, we need to decide what is valid for the param to be
used on.  It can currently only be used on an arithmetic sequence as
in

  for param i in 1..4 do

but perhaps can be generalized.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5658 3a8e244f-b0f2-452b-bcba-4c88e055c3ca