Implemented type select statements via overloaded nested functions. Whether this is the way we want to go (I think so), it didn't take long to get it working, and it is at least correct.  Since the nested functions are inlined, it is equivalent in the end.
There is an error in analysis's handling of nested functions.  The
interpreter passes the two futures checked in here.

I implemented this construct as

  type select ... {
    when ... do ...
    ...

breaking typeselect into two words.  Note this is different from the
specification and should likely be changed to a single word.  However,
one nice thing about the two words is that we eliminate a keyword..

The test

  test/trivial/deitz/test_type_select1-handcode.chpl

illustrates the nested function bug in analysis without using a type
select statement.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5678 3a8e244f-b0f2-452b-bcba-4c88e055c3ca