Added bounds checking to arithmetic domains.
The error message can be improved in the future.  I use exit because
halt prints out the full path of the file which doesn't work in the
test suite without special flags.  The error message prints out the
out-of-bounds index, but should print out the name of the array (needs
to be stored in the Array class or passed to the access function) and
the place in the file where this happens.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5681 3a8e244f-b0f2-452b-bcba-4c88e055c3ca