Implemented initial functionality to build a scalar promotion wrapper where necessary.  It is called only from in the context of the interpreter at this time.
It uses Type::scalarPromotionType which is set in arrays and sequences
and says that this type can promote.

The wrapper is built in FnSymbol::promotion_wrapper and it takes a map
from formal ArgSymbols to TypeSymbols.

I don't think we want to promote the = function to handle assignment
to sequences, arrays, etc., but I think there are other ways to handle
this too.

This gets test/arrays/deitz/test_scalar_promote1.chpl and
test/arrays/deitz/test_scalar_promote2.chpl working correctly in the
interpreter using the --scalar-promotion and -i flags.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5825 3a8e244f-b0f2-452b-bcba-4c88e055c3ca