Added a bunch of tests related to the dispatch thread of the past 24 hours.
dispatch.chpl is the original test I mailed about.  I
made it a future as a placeholder for generating a
better error message and/or adding support for a more
C++/Java-like interpretation as we discussed (possibly
via a compiler flag, possibly as a default to be overridden
via a compiler flag.

dispatch-withnil is the suggestion John made of supplying
a nilType version of the function to resolve the ambiguity.
It works as he suggested.

dispatch-nevernil is the version I asked about in the meeting
where the variable is constructed by default, avoiding the
chance that it will be assigned nil if the user doesn't do
so explicitly.  It works as John predicted.

dispatch-nevernil2 are some variations where the member is
non-initialized by default but is initialized in the constructor.
Half of these work.  The other half seem to die due to problems
invoking a constructor as part of a named argument list.  I
think this is allowed in the language, so futurized them.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5854 3a8e244f-b0f2-452b-bcba-4c88e055c3ca