Add a test for a nested function which subsumes a higher level function.
This test fails because we currently don't support declared subtyping
for records (at least with the familiar syntax).  I seem to remember that
we decided to get rid of structural subtyping for records and add
in declared subtyping, so I think this test should work.
Assigned to Steve on the language issue.  After that
we can talk about whether or no this subsumption nesting issue is
a visibility issue or a dispatch issue (we could handle it in either
place).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5862 3a8e244f-b0f2-452b-bcba-4c88e055c3ca