As discussed at yesterday's meeting added a .future capability for the testing system in interpreted mode using files with the extension .ifuture.  Thus, if a test has a .ifuture file, it will be handled as usual in the testing system under a normal run, but handled as a future test when running the testing system in interpreter mode (using -i or --interpret).
Recall that normal .future tests are currently skipped over when
running the testing system in interpreter mode in order to save
time (under the assumption that if it's not working in normal
mode, it's less likely to work in interpreter mode).

Added .ifuture files for the two largest classes of interpreter
failures: terminating the program early (via halt, assert, or
exit) and using fscanf.  Both of these should be cleared up as
Shannon finishes her varargs versions of these functions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@5933 3a8e244f-b0f2-452b-bcba-4c88e055c3ca