Note: new regressions:
  [Error matching compiler output for types/tuple/bradc/writeTuple]
  [Error matching compiler output for types/tuple/deitz/test_tuples4]

Added support for a _copy function so that variable declarations with
only initialization clauses are now translated as

  var x = y;   ==>   var x; x = _copy(y);

The _copy function is specialized for records.  Default copy
constructors for other types are in _chpl_base.chpl

Removed mechanism for a variable number of fields in a record.  The
implementation of tuples is now based on pragmas that effect during
instantiation.

Fixed a bug in instantiation regarding the caching of instantiated
functions.  The problem was that we would cache functions after they
were instantiated even if the where clause evaluated to false and
would therefore invalidate the instantiated function.  Then a second
attempt to instantiate the same function in the same way would simply
grab the cached copy.  The fix: added a mechanism to remove the cached
copy if the instantiated function is invalidated.  Note that it was
not possible to insert into the cache later (after where clause
evaluation) due to the problem of running into an infinite loop if the
body of the instantiated function during normalization wished to
instantiate the same function in the same way.  The cache catches this
and simply returns the same function.

Moved Symbol::variableExpr to ArgSymbol::variableExpr since it now
only applies to ArgSymbols.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6167 3a8e244f-b0f2-452b-bcba-4c88e055c3ca