Removed "_" unspecified.  Removed support for definition of variables that have no type or initialization expression.  The variable gUnspecified and type dtUnspecified should no longer appear in an AST. They can be removed when this is removed from analysis.
Fixed many tests that did not initialize a field in a class or record
in the constructor call of that class or record if the field had no
type or initialization expression.  Fixed the tests that did something
like

  class C {
    var x;
  }

  var c = C();
  c.x = 2;

by changing them to

  class C {
    var x;
  }
  var c = C(2);

The error message the user gets for this issue is vague and should be
updated.  It is a high-level error message.

Changed reductions to currently require an elt_type so to reduce an
array of integers, one must write

  max(int) reduce A

We can certainly get around this by passing the first element in the
sequence to the reduction class, but that won't work for an empty
sequence.

Cleaned up sequence append and prepend code to initialize the field.
This failed when the field was a tuple because the default constructor
was called and the fields have no type or initialization.  I think
this is an interesting issue for tuples.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6195 3a8e244f-b0f2-452b-bcba-4c88e055c3ca