Removed usage of FnSymbol::typeBinding from many places in the compiler most often by replacing typeBinding->definition with _this->type, but occasionally simplifying.  This is in preparation for supporting bound functions on type expressions.
Added support for bound functions on types that cannot be specified
with a single identifier, e.g., int(32), but are type expressions.
For the time being, for parsing purposes, the type is required to be
in parentheses as in

  fun (int(32)).print() { ...

This work furthers Wayne's work on making the complex type a primitive
type and supporting bit widths on types.  It is also part of my effort
to simplify the normalization/cleanup/scope resolution code.

Fixed FnSymbol constructor to allocate the formals alist.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6226 3a8e244f-b0f2-452b-bcba-4c88e055c3ca