Removed all usage of the INIT primitive from the AST.  References to it are still in the compiler; to be removed next.  Simplified DefExpr transform.
Removed references to dtUnspecified.  Left gUnspecified which is used
in the interpreter for implementing closures.

Added _init function.  It returns the initial value of a type and is
used by the compiler for when it does not know the type before
analysis runs.  For example, "var x : y.type" could be translated into
"_move(x, _init(y))."

The _init functions is treated similarly to the _copy function for
tuples, it is rebuilt during instantiation.

Fixed --no-header-errors to apply to the printing of candidate
functions as well.

Added a hack to param init so that the type is ignored on params that
also have an initialization expression.  This misses a necessary
coercion in the case of "param x : float = 4" and I added a test case
to mark this.  I did this because the param folding currently assumes
there is only one move to a parameter.

Added _init functions for the primitive types and objects, _init
functions for records are automatically generated.  There is not yet
an _init function for the primitive types that have different bit
widths.  It is only necessary, however, for the case where the type is
unknown.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6254 3a8e244f-b0f2-452b-bcba-4c88e055c3ca