Added support for indexing into an array with a domain.  This works for setting values in the array as in
  A(InnerD) = 0.0;

but fails for reading values from an array as in

  writeln(A(InnerD));

This works in the interpreter but fails in analysis with regards to
handling nested functions.  I filed a future for John and will also
look into it further.

The result of reading an array indexed by a domain is an array rather
than a sequence.  We should discuss the semantics for when arrayness
can be maintained without lowering to sequence semantics.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6331 3a8e244f-b0f2-452b-bcba-4c88e055c3ca