Added a new Jacobi future based on my Jacobi implementation.  It works in the interpreter though I did not run it to completion.  There is an analysis failure in the compiler.
To get this to work, I added an overloaded + function at the top of
the program that overloads addition on the two types of tuples thereby
allowing the index from iterating over a domain to be added to the
stencil vector of tuples.

I implemented this just to see how close we were.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6335 3a8e244f-b0f2-452b-bcba-4c88e055c3ca