Removed future for test that indexes into an array with a domain that John fixed.
Removed future for Brad's test of assigning one array the negative of
another as in A = -B which started working now that scalar promotion
on functions that return values works.

Added a test to assign the result of adding two arrays to a third
array as in C = A + B.  This works with zipper iteration.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6363 3a8e244f-b0f2-452b-bcba-4c88e055c3ca