We sometimes need to create a list of Stmts rather than a single Stmt when we parse a Chapel statement.  To this end, the Chapel grammar was modified so that a statement is parsed as an AList<Stmt>.  However this makes it difficult to determine the extent of a single Chapel statement.
Added a new type of BlockStmt called BLOCK_SCOPELESS that is like a
BlockStmt but that has no scope associated with it.  It is thus
organizational only.

Modified the Chapel grammar so that Chapel statements are parsed as
scopeless BlockStmts.

This is necessary to support the transformation on cobegin blocks that
seek to parallelize at the Chapel statement level.  Removed future for
a test of a while loop in a cobegin block.

Added to the cleanup pass functionality to flatten all scopeless
blocks.  To discuss: Should we flatten these blocks or leave them in
as representative of where the Chapel statements are.  I can imagine
they may be useful for looking at the code and/or debugging.  Not
removing the scopeless blocks can be accomplished by not calling the
flatten_scopeless_blocks function in cleanup.  This has been tested
and works on Jacobi.

Reassociated the uses vector with BaseAST rather than with SymScope
since the SymScope is wiped out when AST is removed and uses need to
be maintained.

Added a new variant on collect_asts called collect_top_asts that does
not recurse into a symbol (other than the initial symbol passed to it
if that is a symbol).  This could be useful to collect the asts in a
function without also collecting the asts in nested functions.  It is
used now in the ClassType::addDeclarations to determine primary
methods, fields, and types in a class.  This makes this code more
robust necessary to handle the additional BlockStmts inserted with
this checkin.  I modified a bunch of other fragile code though some of
it remains fragile.

Made the new traversal mechanism work when passed a list of
statements.  It still complains with other lists.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6366 3a8e244f-b0f2-452b-bcba-4c88e055c3ca